Fixed a few warnings.
Rev: src/modules/HTTPLoop/accept_and_parse.c:1.17
Rev: src/modules/HTTPLoop/cache.c:1.13
Rev: src/modules/HTTPLoop/cache.h:1.5
Rev: src/modules/HTTPLoop/log.c:1.8
Rev: src/modules/HTTPLoop/requestobject.c:1.13
Rev: src/modules/HTTPLoop/util.c:1.6
Rev: src/modules/Image/encodings/gif.c:1.57
Rev: src/modules/Image/encodings/ilbm.c:1.19
Rev: src/modules/Image/encodings/ras.c:1.11
Rev: src/modules/Image/encodings/x.c:1.31
Rev: src/modules/Image/font.c:1.63

+++ src/modules/HTTPLoop/accept_and_parse.c
@@ -264,7 +264,7 @@ static int parse(struct args *arg)
 			      arg->res.host, arg->res.host_len,
 			      arg->cache,0, NULL, NULL)) && ce->data)
 	{
-	  int len = WRITE(arg->fd, ce->data->str, ce->data->len);
+	  ptrdiff_t len = WRITE(arg->fd, ce->data->str, ce->data->len);
 	  LOG(len, arg, atoi(ce->data->str+MY_MIN(ce->data->len, 9))); 
 	  simple_aap_free_cache_entry( arg->cache, ce );
 	  /* if keepalive... */
@@ -287,7 +287,7 @@ static int parse(struct args *arg)
 void aap_handle_connection(struct args *arg)
 {
   char *buffer, *p, *tmp;
-  int pos, buffer_len;
+  ptrdiff_t pos, buffer_len;
 #ifdef HAVE_TIMEOUTS
   int *timeout = NULL;
 #endif
